add validation login override spot option in deployer.go and add test case to deployer_test.go
